Prefer iTerm graphics protocol when enabled
